+++ spec/integration/google_tag_manager_integration_spec.rb
@@ -11,8 +11,8 @@ RSpec.describe "Google Tag Manager Integration" do
   subject { page }
 
   it "embeds the script tag with tracking event from the controller action" do
-    expect(page.find("body")).to have_content('GTM-ABCDEF')
-    expect(page.find("body")).to have_content('\'click\': \'X\', \'price\': \'10\'')
+    expect(page.find("body")).to have_content 'GTM-ABCDEF'
+    expect(page.find("body")).to have_content "dataLayer.push( {\"click\":\"X\",\"price\":10}, {\"transactionProducts\":[{\"sku\":\"DD44\",\"name\":\"T-shirt\"},{\"sku\":\"DD66\",\"name\":\"Jeans\"}]} );"
   end
 
 end