+++ spec/support/metal_controller.rb
@@ -42,8 +42,8 @@ class MetalController < ActionController::Metal
 
   def google_tag_manager
     tracker do |t|
-      t.google_tag_manager :push, { name: 'click', value: 'X' }
-      t.google_tag_manager :push, { name: 'price', value: '10' }
+      t.google_tag_manager :push, { click: 'X', price: 10 }
+      t.google_tag_manager :push, transactionProducts: [{ sku: 'DD44', name: 'T-shirt' }, { sku: 'DD66', name: 'Jeans' }]
     end
     render "metal/index"
   end